Some build procedure improvements; application build rules can be acquired from config (not necessarely hardcoded in testsuite/thirdparty/_local).

+++ build.sh
@@ -12,6 +12,15 @@ function usage() {
     fatal "command line parsing results in no further action."
 }
 
+function check_prereq_exists() {
+    pkg="$1"
+    if [ -e thirdparty/${pkg}.bin -a -e thirdparty/${pkg}.src ]; then 
+        return 0; 
+    else
+        return 1;
+    fi
+}
+
 set -ue
 
 source thirdparty/dbscripts/base.inc
@@ -38,40 +47,55 @@ if [ ! -e "$basedir" ]; then
     du_gitclone_recursive "confs" "$url" "HEAD" "du"
 fi
 
-hwdir=$basedir/$app/$testmodule/$hwconf
-
+appdir="$basedir/$app/$testmodule"
+hwdir="$basedir/$app/$testmodule/$hwconf"
 [ ! -d "$hwdir" ] && fatal "can't find configuration: $hwconf in config directory. Tried to access directory: $hwdir."
 
 [ -f "$hwdir"/env.sh ] && cp "$hwdir"/env.sh . || fatal "no env.sh file in $hwdir."
-dir="$hwdir/$suite_name"
+suite_dir="$hwdir/$suite_name"
 
-[ ! -d "$dir" ] && fatal "can't find suite_name: $suite_name in config directory. Tried to access directory: $dir."
+[ ! -d "$suite_dir" ] && fatal "can't find suite_name: $suite_name in config directory. Tried to access directory: $suite_dir."
 [ -e "$app.conf" ] && rm "$app.conf"
+ln -s "$suite_dir" "$app.conf"
 
-ln -s "$dir" "$app.conf"
+[ -e thirdparty/_local/conf.inc -o -L thirdparty/_local/conf.inc ] && rm -f thirdparty/_local/conf.inc
+if [ -e thirdparty/_local/$app.inc ]; then
+    ln -s $app.inc thirdparty/_local/conf.inc
+else
+    [ -e "$appdir/build.inc" ] || fatal "can't find build script for application: $app. Tried to access file: $appdir/build.inc."
+    ln -s $PWD/$appdir/build.inc thirdparty/_local/conf.inc
+fi
 
-[ -r thirdparty/_local/conf.inc ] && rm -f thirdparty/_local/conf.inc
-
-ln -s $app.inc thirdparty/_local/conf.inc
 
 export TESTSUITE_SUITE_NAME="$suite_name"
 
-s=$(ls -1d thirdparty/*-*.src 2>/dev/null | wc -l)
 
 if [ -z "$dont_always_rebuild" ]; then
-
-if [ "$s" != "0" ]; then
-    rm -rf thirdparty/*-*.src thirdparty/sandbox
-fi
-cd thirdparty
-./dnb.sh
-cd ..
-
+    TESTSUITE_PACKAGES_EXPR=$(grep 'TESTSUITE_PACKAGES=' thirdparty/_local/conf.inc)
+    PREREQUISITES=""
+    ALL_PREREQ=$(check_prereq_exists "argsparser" && check_prereq_exists "daemonize" && check_prereq_exists "psubmit" && check_prereq_exists "yaml-cpp" && check_prereq_exists "massivetests" && echo OK || true)
+    if [ -z "$TESTSUITE_PACKAGES_EXPR" -o "$ALL_PREREQ" != "OK" ]; then
+        rm -rf thirdparty/*-*.src thirdparty/*.bin thirdparty/sandbox
+        cd thirdparty
+        ./dnb.sh
+        cd ..
+    else
+        eval $TESTSUITE_PACKAGES_EXPR
+        for pkg in $TESTSUITE_PACKAGES; do
+            rm -rf thirdparty/${pkg}-*.src
+        done
+        rm -rf thirdparty/sandbox
+        cd thirdparty
+        for pkg in $TESTSUITE_PACKAGES; do
+            ./dnb.sh ${pkg}:dub
+        done
+        ./dnb.sh :i
+        cd ..
+    fi
 else
-
-if [ "$s" == "0" ]; then dnbmode=""; else dnbmode=":bi"; fi
-cd thirdparty
-./is_rebuild_required.sh && ./dnb.sh "$dnbmode" || ./dnb.sh massivetests:i
-cd ..
-
+    s=$(ls -1d thirdparty/*-*.src 2>/dev/null | wc -l)
+    if [ "$s" == "0" ]; then dnbmode=""; else dnbmode=":bi"; fi
+    cd thirdparty
+    ./is_rebuild_required.sh && ./dnb.sh "$dnbmode" || ./dnb.sh massivetests:i
+    cd ..
 fi
